Fixed stretch/resize.
Resizing respects minwidth/height by recursing down through the stretched object to see if any of its contents would resize beyond minimum width. If so, it changes the resize amount to the amount that makes it exactly enough to go to minimum, and passes that back up.
Also, to aid debugging, all widgets now have an is_a field.